pyproject.toml
[project]
name = "gitexplorer"
version = "0.1.0"
description = "A PyQt6 Git history explorer — browse file history with syntax-highlighted diffs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "gitexplorer contributors" }]
keywords = ["git", "history", "diff", "pyqt6", "gui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "PyQt6>=6.6.0",
    "gitpython>=3.1.40",
    "pygments>=2.17.0",
]

[project.scripts]
gitexplorer = "gitexplorer.__main__:main"

[project.urls]
Homepage = "https://github.com/SvenBroeckling/gitexplorer"
Repository = "https://github.com/SvenBroeckling/gitexplorer"
Issues = "https://github.com/SvenBroeckling/gitexplorer/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gitexplorer"]
